--- app/src/main/java/com/treasurehunt/data/remote/ImageRepository.kt
@@ -1,8 +1,0 @@
-package com.treasurehunt.data.remote
-
-import com.treasurehunt.data.remote.model.ImageDTO
-
-interface ImageRepository {
-
-    suspend fun insertImage(imageDTO: ImageDTO): String
-}
