--- app/src/main/java/com/treasurehunt/data/remote/ImageRepositoryImpl.kt
@@ -1,11 +1,0 @@
-package com.treasurehunt.data.remote
-
-import com.treasurehunt.data.remote.model.ImageDTO
-import com.treasurehunt.data.remote.model.ImageRemoteDataSource
-
-class ImageRepositoryImpl(
-    val imageRemoteDataSource: ImageRemoteDataSource
-) : ImageRepository {
-    override suspend fun insertImage(imageDTO: ImageDTO): String =
-        imageRemoteDataSource.insert(imageDTO)
-}
